+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+**v0.1.2 - 01-22-2026**
+---
+
+- Looses keyboard control if e-mail content was clicked
+- Autofocus on first message when switched to new folder
+- Disable focus on conversation viewer when links are clicked
+
+
 **v0.1.1 - 01-19-2026**
 ---